+++ packages/ap2/jest.config.js
@@ -1,0 +1,12 @@
+export default {
+  bail: true,
+  clearMocks: true,
+  testEnvironment: 'node',
+  transform: {
+    '^.+\\.(ts|js)$': ['babel-jest', { rootMode: 'upward' }],
+  },
+  transformIgnorePatterns: [
+    '/node_modules/(?!@stablelib|uuid|@ar-agents/ap2|jose)',
+  ],
+  verbose: true,
+};