+++ packages/crypto-utils/src/vc/index.js
@@ -1,3 +1,5 @@
 export * from './crypto';
 export * from './jws';
 export * from './jws-signer';
+export * from './jwk';
+export * from './sd-jwt';